final-pjt-back/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
binary==1.0.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.11
django-allauth==0.61.1
django-appconf==1.0.6
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-imagekit==5.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jiter==0.7.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
logging==0.4.9.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
mysqlclient==2.2.6
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.55.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.8
phonenumbers==8.13.50
pilkit==3.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pywin32==307
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
yfinance==0.2.50
zipp==3.19.2
